app adjustments
All apps: Max file size is upgraded to 10MB  /  Maps now can spiderify when markers are overlapping  /  Tables are presented as DataTables, to save on space
Fish: The MasterList is updated to include A & B factors for calculating biomass from length.
Formatter: Several usability adjustments  /  Biomass is now calculaated based on pre-downloaded data so it is quicker  /  Column name mapping is complemented with more variations  /  Data upload doesn't try to repair column names, it was causing issues with data compatibility.